backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
python-multipart>=0.0.6
passlib>=1.7.4
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.1.0
pypdf>=4.0.0
pdfplumber>=0.10.3
pdf2image>=1.16.3
pytesseract>=0.3.10
Pillow>=10.2.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
email-validator>=2.1.0
jinja2>=3.1.0
boto3>=1.34.0
